Revert "- signin, signup and recovery pages made separate at components"
