+++ src/components/passRecovery.js
@@ -1,53 +1,15 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
+import { validatePassReset } from "./validateInfo";
+import useForm from "./useForm";
 import { Avatar, Button, CssBaseline, TextField, Grid, Box, Container, Typography } from "@material-ui/core";
 import { useStyles } from "./usesstyles";
 import LockOpenIcon from "@material-ui/icons/LockOpen";
 import { Link } from "react-router-dom";
-// import { validatePassReset } from "./validateInfo";
-// import useForm from "./useForm";
 
-const PassworResetComponent = ({ submitForm }) => {
-  // PUSSWORD RESET IN VALIDATION
-  const validatePassReset = (values) => {
-    let errors = {};
-    if (!values.email) {
-      errors.email = "Email required";
-    } else if (!/\S+@\S+\.\S+/.test(values.email)) {
-      errors.email = "Email address is invalid";
-    }
-    return errors;
-  };
-  //FORM SUBMID FUNCTION
-  const useForm = (callback, validate) => {
-    const [values, setValues] = useState({
-      email: "",
-    });
-    const [errors, setErrors] = useState({});
-    const [isSubmitting, setIsSubmitting] = useState(false);
-    const handleChange = (e) => {
-      const { name, value } = e.target;
-      setValues({
-        ...values,
-        [name]: value,
-      });
-    };
-    const handleSubmit = (e) => {
-      e.preventDefault();
-      setErrors(validate(values));
-      setIsSubmitting(true);
-    };
-    useEffect(() => {
-      if (Object.keys(errors).length === 0 && isSubmitting) {
-        console.log("Your new password will be sent on email");
-      } else if (Object.keys(errors).length !== 0 && isSubmitting) {
-        console.log("Nop, you have fill up all forms");
-      }
-    }, [isSubmitting, errors]);
-    return { handleChange, handleSubmit, values, errors };
-  };
-  // call functions FORM SUBMID & VALIDATION
+const PassworReset = ({ submitForm }) => {
   const { handleChange, handleSubmit, values, errors } = useForm(submitForm, validatePassReset);
   const classes = useStyles();
+
   return (
     <div className="login">
       <Container component="main" maxWidth="xs">
@@ -98,4 +60,4 @@ const PassworResetComponent = ({ submitForm }) => {
   );
 };
 
-export default PassworResetComponent;
+export default PassworReset;
