--- src/components/passResetPage.js
@@ -1,12 +1,0 @@
-import React from "react";
-import PassworResetComponent from "./passResetComponent";
-
-const SignUpPage = () => {
-  return (
-    <>
-      <PassworResetComponent />
-    </>
-  );
-};
-
-export default SignUpPage;
