--- src/components/signInPage.js
@@ -1,12 +1,0 @@
-import React from "react";
-import SignInComponent from "./signInComponent";
-
-const SignInPage = () => {
-  return (
-    <>
-      <SignInComponent />
-    </>
-  );
-};
-
-export default SignInPage;
